# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Anthology CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: List email templates
      parameters:
        - name: audience
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailTemplate"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailTemplate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /email/send:
    post:
      operationId: sendEmail
      tags: [email]
      summary: Send an email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmail"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messageId:
                    type: string
...
components:
...
  schemas:
...
    EmailTemplate:
      type: object
      required: [id, name, audience, subject, body, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        audience:
          type: string
        subject:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateEmailTemplate:
      type: object
      required: [name, audience, subject, body]
      properties:
        name:
          type: string
        audience:
          type: string
        subject:
          type: string
        body:
          type: string
...
    SendEmail:
      type: object
      required: [to, subject, body]
      properties:
        to:
          type: string
        subject:
          type: string
        body:
          type: string
        leadId:
          type: integer
        contactId:
          type: integer
        addToCalendar:
          type: boolean
...